Simplify InstUpdateInstaller tests
