+++ test/support/fake_registration.rb
@@ -1,0 +1,11 @@
+# Registration::Storage::InstallationOptions fake
+class FakeInstallationOptions
+  include Singleton
+  attr_accessor :custom_url
+end
+
+# Registration::Storage::Config fake
+class FakeRegConfig
+  include Singleton
+  def import(_args); end
+end
